ENH: Add the class NonDicomFileClassifierFillerStep to create the basic hierarchy for non-DICOM files.
This class replaces the old MHDFileClassifierStep and supports all the
image types supported by vtkImageReader2 subclasses, e.g. MetaImage,
TIFF, PNG, JPEG, BMP and several more.

It tries to support both one series in one file and one series in
several files (i.e. a MetaImage vs a collection of JPEGs) at the same
time. To do so it guesses that several files will go to the same series
if they are single-frame, are in the same directory and have the same
name ending with a different number. This may need adjustments in the
future.